cloud/src/commons_cloud/providers/gcp/gcs.py
"""Google Cloud Storage implementation."""

from typing import AsyncIterator, Dict, List, Optional, Union
from pathlib import Path
import io

from google.cloud import storage
from commons_core.logging import get_logger
from ...abstractions.storage import (
    StorageProvider,
    StorageObject,
    UploadOptions,
    SignedUrlOptions,
)

logger = get_logger(__name__)


class GCSStorage(StorageProvider):
    """Google Cloud Storage provider."""
    
    def __init__(
        self,
        bucket: str,
        project_id: Optional[str] = None,
        credentials_path: Optional[str] = None,
    ) -> None:
        super().__init__(bucket)
        # Implementation would initialize GCS client
        self.project_id = project_id
        self.credentials_path = credentials_path
        
    # Implementation placeholder - full implementation would follow
    async def upload(
        self,
        key: str,
        data: Union[bytes, io.IOBase, Path],
        options: Optional[UploadOptions] = None,
    ) -> StorageObject:
        """Upload an object to GCS."""
        # Full implementation would go here
        return StorageObject(key=key, size=0, last_modified=None, etag="")