cloud/pyproject.toml
[project]
name = "commons-cloud"
version = "0.1.0"
description = "Cloud provider abstractions for Python Commons"
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Python Commons Team", email = "team@pythoncommons.org"}
]
dependencies = [
    "commons-core",
    "aioboto3>=12.0.0",
    "boto3>=1.28.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-compute>=1.14.0",
    "google-cloud-secret-manager>=2.16.0",
    "azure-storage-blob>=12.19.0",
    "azure-mgmt-compute>=30.3.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-identity>=1.14.0",
]

[tool.uv.sources]
commons-core = { workspace = true }

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "moto[s3,ec2,secretsmanager]>=4.2.0",
]
aws = ["boto3>=1.28.0", "aioboto3>=12.0.0"]
gcp = [
    "google-cloud-storage>=2.10.0",
    "google-cloud-compute>=1.14.0",
    "google-cloud-secret-manager>=2.16.0",
]
azure = [
    "azure-storage-blob>=12.19.0",
    "azure-mgmt-compute>=30.3.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-identity>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/commons_cloud"]